requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
opencv-python==4.8.1.78

